Little bit of polish
- Introduced a new ICardHolder interface
- Created an empty CardDeck class for later functionality
- Moved card scripts into a sub folder
- Updated the several card scripts with clean up tweaks
