+++ Assets/Scenes/Cards and Input.unity
@@ -400,14 +400,6 @@ PrefabInstance:
     m_TransformParent: {fileID: 944109912}
     m_Modifications:
     - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
-      propertyPath: m_Pivot.x
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
-      propertyPath: m_Pivot.y
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
       propertyPath: m_RootOrder
       value: -1
       objectReference: {fileID: 0}
@@ -610,14 +602,6 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 944109912}
     m_Modifications:
-    - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
-      propertyPath: m_Pivot.x
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
-      propertyPath: m_Pivot.y
-      value: 0.5
-      objectReference: {fileID: 0}
     - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
       propertyPath: m_RootOrder
       value: -1
@@ -833,8 +817,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _inputs: {fileID: 568748454}
   _draggables: {fileID: 44541900}
-  offsetPosition: {x: 45, y: -25}
-  offsetRotation: {x: 0, y: 0, z: -15}
+  _offsetPosition: {x: 0.2, y: 0.8}
+  _offsetRotation: {x: 0, y: 0, z: -15}
 --- !u!1 &1394222707
 GameObject:
   m_ObjectHideFlags: 0
@@ -975,14 +959,6 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 944109912}
     m_Modifications:
-    - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
-      propertyPath: m_Pivot.x
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
-      propertyPath: m_Pivot.y
-      value: 0.5
-      objectReference: {fileID: 0}
     - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
       propertyPath: m_RootOrder
       value: -1
@@ -1356,14 +1332,6 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 944109912}
     m_Modifications:
-    - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
-      propertyPath: m_Pivot.x
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
-      propertyPath: m_Pivot.y
-      value: 0.5
-      objectReference: {fileID: 0}
     - target: {fileID: 6662708026035231147, guid: 233ea7c07dd563f42a403045f00ea88f, type: 3}
       propertyPath: m_RootOrder
       value: -1
