+++ Assets/Scripts/Card System.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06ef1d549a75eb545adfc2d6b317e835
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
