+++ Assets/Scripts/Card System/CardDeck.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9b691b539d2af724689381bc5eed1b0c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
