+++ Assets/Scripts/Card System/CardManager.cs
@@ -11,10 +11,11 @@ public class CardManager : MonoBehaviour
     private InputAction _mousePosition;
 
     [SerializeField] private Transform _draggables;
+    // This is obselete maybe now lol
     [Tooltip("The offset from the center when picking up an dragging a card. This is purely cosmetic")]
-    [SerializeField] private Vector2 offsetPosition;
+    [SerializeField] private Vector2 _offsetPosition = new Vector3(0.2f, 0.8f);
     [Tooltip("The rotation offset to make the card look angled when being picked up.")]
-    [SerializeField] private Vector3 offsetRotation;
+    [SerializeField] private Vector3 _offsetRotation = Vector3.zero;
 
     #region Unity Callbacks
 
@@ -36,22 +37,22 @@ public class CardManager : MonoBehaviour
         if (CarriedCard == null)
             return;
 
-        CarriedCard.transform.position = _mousePosition.ReadValue<Vector2>() + offsetPosition;
+        CarriedCard.transform.position = _mousePosition.ReadValue<Vector2>();
     }
-
     
     public void SetCarriedCard(Card card)
     {
         // The CardManager should handle swapping cards
         if (CarriedCard != null && card != null)
-            card.ActiveSlot.SetCard(CarriedCard);
+            card.ActiveSlot.SetHeldCard(CarriedCard);         
 
         // Carried card settings
         CarriedCard = card;
         if (card != null)
         {
             card.transform.SetParent(_draggables);
-            card.transform.Rotate(offsetRotation, Space.Self);
+            card.transform.Rotate(_offsetRotation, Space.Self);
+            (card.transform as RectTransform).pivot = _offsetPosition;
         }
     }
 }
