+++ Assets/Scripts/Card System/ICardHolder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4713a41a80ac69f49b69aa105ad476fb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
